backend/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
bidict==0.23.1
boto3==1.40.6
botocore==1.40.6
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
cryptography==45.0.6
distlib==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
filelock==3.20.0
h11==0.16.0
httptools==0.6.4
identify==2.6.15
idna==3.10
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.18.2
mypy_extensions==1.1.0
nodeenv==1.9.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pre_commit==4.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
rsa==4.9.1
ruff==0.14.2
s3transfer==0.13.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.35.3
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
